refactor: move settings component functions to useSettings custom hook
